+++ src/control-plane-services/instance-cluster-management/icms-core/src/test/java/com/nvidia/icms/outbound/nats/NatsMessageSenderClientIntegrationTest.java
@@ -86,7 +86,7 @@ class NatsMessageSenderClientIntegrationTest extends IntegrationTest {
 
         // Spring lifecycle callbacks do not run for manually constructed test objects.
         NatsStreamManager natsStreamManager = new NatsStreamManager(
-                new NatsResourceService(fixedNatsPool.borrowConnection()),
+                fixedNatsPool.borrowConnection(),
                 natsConfigurationProperties);
         natsStreamManager.init();
